3.1.0-ubi: add UBI 8 container definition
Adds a Dockerfile and supporting files to run CouchDB 3.1.0
on UBI 8. This is largely a clone of the 2.3.1-UBI assets, updated
where required, to support CouchDB 3.x.

Specifically, the default entrypoint now requires an admin username
and password to be supplied.

Since it installs the community-maintained rpm, the container
uses SpiderMonkey 60 as the JavaScript runtime for CouchDB.